Fix bug causing 29-bit format IDs to be truncated
Previously, the format of the written CAN message was not checked and the ID would be truncated to the 11-bit short format. This commit fixes this behavior and sets the XTD bit of the TCAN TX header as appropriate based on the given CAN message format.

Also introduces tcan trace group and result code checking for write operations.